requirements.txt
git+https://github.com/pyrogram/pyrogram@master
TgCrypto==1.2.2
py-tgcalls==0.8.0
python-dotenv
youtube_dl
youtube_search_python
youtube-dl
pytube
youtube_search
requests
sqlalchemy==1.3.23
psycopg2-binary 
ffmpeg-python
moviepy